Propagator: hide PropagatorError implementation
